db opd err

+++ models/form.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, DataTypes) => {
   var Form = sequelize.define('form', {
     nama: DataTypes.STRING,
     keterangan: DataTypes.STRING,
-    id_opd: {type: DataTypes.INTEGER(9), allowNull: false},
+    id_opd: {type: DataTypes.INTEGER, allowNull: false},
     id_dataset: {type: DataTypes.INTEGER, allowNull: false}
   }, {});
   Form.associate = function(models) {
